Feature/personal settings

+++ src/lib/components/chat/Settings/Personalization/ManageModal.svelte
@@ -12,6 +12,11 @@
 	import { error } from '@sveltejs/kit';
 	import EditMemoryModal from './EditMemoryModal.svelte';
 	import localizedFormat from 'dayjs/plugin/localizedFormat';
+	import EllipsisHorizontal from '$lib/components/icons/EllipsisHorizontal.svelte';
+	import MemoriesMenu from './MemoriesMenu.svelte';
+	import AddEditMemory from './AddEditMemory.svelte';
+	import DeleteConfirmDialog from '$lib/components/common/ConfirmDialog.svelte';
+	import DOMPurify from 'dompurify';
 
 	const i18n = getContext('i18n');
 	dayjs.extend(localizedFormat);
@@ -32,130 +37,83 @@
 			loading = false;
 		})();
 	}
+	let showDeleteConfirm = false;
 </script>
 
-<Modal size="xl" bind:show>
+<Modal size="md-plus" bind:show className="dark:bg-customGray-800 rounded-2xl">
 	<div>
-		<div class=" flex justify-between dark:text-gray-300 px-5 pt-4 pb-1">
-			<div class=" text-lg font-medium self-center">{$i18n.t('Memory')}</div>
-			<button
-				class="self-center"
-				on:click={() => {
-					show = false;
-				}}
+		<div class="px-7">
+			<div
+				class=" flex justify-between dark:text-white pt-5 pb-4 border-b dark:border-customGray-700"
 			>
-				<svg
-					xmlns="http://www.w3.org/2000/svg"
-					viewBox="0 0 20 20"
-					fill="currentColor"
-					class="w-5 h-5"
+				<div class="self-center">{$i18n.t('Memory')}</div>
+				<button
+					class="self-center"
+					on:click={() => {
+						show = false;
+					}}
 				>
-					<path
-						d="M6.28 5.22a.75.75 0 00-1.06 1.06L8.94 10l-3.72 3.72a.75.75 0 101.06 1.06L10 11.06l3.72 3.72a.75.75 0 101.06-1.06L11.06 10l3.72-3.72a.75.75 0 00-1.06-1.06L10 8.94 6.28 5.22z"
-					/>
-				</svg>
-			</button>
+					<svg
+						xmlns="http://www.w3.org/2000/svg"
+						viewBox="0 0 20 20"
+						fill="currentColor"
+						class="w-5 h-5"
+					>
+						<path
+							d="M6.28 5.22a.75.75 0 00-1.06 1.06L8.94 10l-3.72 3.72a.75.75 0 101.06 1.06L10 11.06l3.72 3.72a.75.75 0 101.06-1.06L11.06 10l3.72-3.72a.75.75 0 00-1.06-1.06L10 8.94 6.28 5.22z"
+						/>
+					</svg>
+				</button>
+			</div>
 		</div>
 
 		<div class="flex flex-col w-full px-5 pb-5 dark:text-gray-200">
 			<div
-				class=" flex flex-col w-full sm:flex-row sm:justify-center sm:space-x-6 h-[28rem] max-h-screen outline outline-1 rounded-xl outline-gray-100 dark:outline-gray-800 mb-4 mt-1"
+				class=" flex flex-col w-full sm:flex-row sm:justify-center sm:space-x-6 h-[20rem] max-h-screen mb-4 mt-1"
 			>
 				{#if memories.length > 0}
 					<div class="text-left text-sm w-full mb-4 overflow-y-scroll">
 						<div class="relative overflow-x-auto">
-							<table class="w-full text-sm text-left text-gray-600 dark:text-gray-400 table-auto">
-								<thead
-									class="text-xs text-gray-700 uppercase bg-transparent dark:text-gray-200 border-b-2 dark:border-gray-800"
+							{#each memories as memory}
+								<div
+									id="memory-{memory.id}"
+									class="flex justify-between items-center group rounded-md w-full dark:hover:bg-customGray-900 py-2 px-3 cursor-pointer"
 								>
-									<tr>
-										<th scope="col" class="px-3 py-2"> {$i18n.t('Name')} </th>
-										<th scope="col" class="px-3 py-2 hidden md:flex">
-											{$i18n.t('Last Modified')}
-										</th>
-										<th scope="col" class="px-3 py-2 text-right" />
-									</tr>
-								</thead>
-								<tbody>
-									{#each memories as memory}
-										<tr class="border-b dark:border-gray-800 items-center">
-											<td class="px-3 py-1">
-												<div class="line-clamp-1">
-													{memory.content}
-												</div>
-											</td>
-											<td class=" px-3 py-1 hidden md:flex h-[2.5rem]">
-												<div class="my-auto whitespace-nowrap">
-													{dayjs(memory.updated_at * 1000).format('LLL')}
-												</div>
-											</td>
-											<td class="px-3 py-1">
-												<div class="flex justify-end w-full">
-													<Tooltip content="Edit">
-														<button
-															class="self-center w-fit text-sm px-2 py-2 hover:bg-black/5 dark:hover:bg-white/5 rounded-xl"
-															on:click={() => {
-																selectedMemory = memory;
-																showEditMemoryModal = true;
-															}}
-														>
-															<svg
-																xmlns="http://www.w3.org/2000/svg"
-																fill="none"
-																viewBox="0 0 24 24"
-																stroke-width="1.5"
-																stroke="currentColor"
-																class="w-4 h-4 s-FoVA_WMOgxUD"
-																><path
-																	stroke-linecap="round"
-																	stroke-linejoin="round"
-																	d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L6.832 19.82a4.5 4.5 0 0 1-1.897 1.13l-2.685.8.8-2.685a4.5 4.5 0 0 1 1.13-1.897L16.863 4.487Zm0 0L19.5 7.125"
-																	class="s-FoVA_WMOgxUD"
-																/></svg
-															>
-														</button>
-													</Tooltip>
+									<div class="line-clamp-1 text-sm dark:text-customGray-100">
+										{memory.content}
+									</div>
+									<div class="invisible group-hover:visible">
+										<MemoriesMenu
+											editHandler={async () => {
+												selectedMemory = memory;
+												// showEditMemoryModal = true;
+											}}
+											deleteHandler={async () => {
+												const res = await deleteMemoryById(localStorage.token, memory.id).catch(
+													(error) => {
+														toast.error(`${error}`);
+														return null;
+													}
+												);
 
-													<Tooltip content="Delete">
-														<button
-															class="self-center w-fit text-sm px-2 py-2 hover:bg-black/5 dark:hover:bg-white/5 rounded-xl"
-															on:click={async () => {
-																const res = await deleteMemoryById(
-																	localStorage.token,
-																	memory.id
-																).catch((error) => {
-																	toast.error(`${error}`);
-																	return null;
-																});
-
-																if (res) {
-																	toast.success($i18n.t('Memory deleted successfully'));
-																	memories = await getMemories(localStorage.token);
-																}
-															}}
-														>
-															<svg
-																xmlns="http://www.w3.org/2000/svg"
-																fill="none"
-																viewBox="0 0 24 24"
-																stroke-width="1.5"
-																stroke="currentColor"
-																class="w-4 h-4"
-															>
-																<path
-																	stroke-linecap="round"
-																	stroke-linejoin="round"
-																	d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0"
-																/>
-															</svg>
-														</button>
-													</Tooltip>
-												</div>
-											</td>
-										</tr>
-									{/each}
-								</tbody>
-							</table>
+												if (res) {
+													toast.success($i18n.t('Memory deleted successfully'));
+													memories = await getMemories(localStorage.token);
+												}
+											}}
+											onClose={() => {}}
+										>
+											<button
+												class="self-center w-fit text-sm px-0.5 h-[21px] dark:text-white dark:hover:text-white hover:bg-black/5 dark:hover:bg-customGray-900 rounded-md"
+												type="button"
+												on:click={(e) => {}}
+											>
+												<EllipsisHorizontal className="size-5" />
+											</button>
+										</MemoriesMenu>
+									</div>
+								</div>
+							{/each}
 						</div>
 					</div>
 				{:else}
@@ -166,33 +124,55 @@
 					</div>
 				{/if}
 			</div>
-			<div class="flex text-sm font-medium gap-1.5">
-				<button
+			<AddEditMemory
+				memory={selectedMemory}
+				on:save={async () => {
+					memories = await getMemories(localStorage.token);
+					selectedMemory = null;
+				}}
+			/>
+			<div class="flex justify-end text-sm font-medium border-t border-customGray-700 pt-5">
+				<!-- <button
 					class=" px-3.5 py-1.5 font-medium hover:bg-black/5 dark:hover:bg-white/5 outline outline-1 outline-gray-300 dark:outline-gray-800 rounded-3xl"
 					on:click={() => {
 						showAddMemoryModal = true;
 					}}>{$i18n.t('Add Memory')}</button
-				>
+				> -->
 				<button
-					class=" px-3.5 py-1.5 font-medium text-red-500 hover:bg-black/5 dark:hover:bg-white/5 outline outline-1 outline-red-300 dark:outline-red-800 rounded-3xl"
+					class=" text-xs h-10 px-4 py-2 transition rounded-lg bg-black hover:bg-gray-900 text-white dark:bg-customGray-900 dark:hover:bg-customGray-950 dark:text-customGray-200 border dark:border-customGray-700 flex justify-center items-center"
 					on:click={async () => {
-						const res = await deleteMemoriesByUserId(localStorage.token).catch((error) => {
-							toast.error(`${error}`);
-							return null;
-						});
-
-						if (res) {
-							toast.success($i18n.t('Memory cleared successfully'));
-							memories = [];
-						}
-					}}>{$i18n.t('Clear memory')}</button
+						if(memories.length < 1) return;
+						showDeleteConfirm = true;
+					}}>{$i18n.t('Delete All Memories')}</button
 				>
 			</div>
 		</div>
 	</div>
 </Modal>
 
-<AddMemoryModal
+<DeleteConfirmDialog
+	bind:show={showDeleteConfirm}
+	title={$i18n.t('Are you sure you want to delete everything?')}
+	on:confirm={async () => {
+		const res = await deleteMemoriesByUserId(localStorage.token).catch((error) => {
+			toast.error(`${error}`);
+			return null;
+		});
+		if (res) {
+			toast.success($i18n.t('Memory cleared successfully'));
+			memories = [];
+		}
+	}}
+	confirmLabel={$i18n.t('Delete All')}
+	>
+	<div class=" text-sm text-gray-700 dark:text-gray-300 flex-1 line-clamp-3">
+		{@html DOMPurify.sanitize(
+			$i18n.t('This action is permanent and cannot be undone. All your data will be lost.')
+		)}
+	</div>
+</DeleteConfirmDialog>
+
+<!-- <AddMemoryModal
 	bind:show={showAddMemoryModal}
 	on:save={async () => {
 		memories = await getMemories(localStorage.token);
@@ -205,4 +185,4 @@
 	on:save={async () => {
 		memories = await getMemories(localStorage.token);
 	}}
-/>
+/> -->
